backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.22.1  # For SQLite async support
alembic==1.13.1

# Supabase
supabase==2.3.4

# HTTP Client
httpx>=0.24,<0.26
aiohttp==3.9.1

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # bcrypt 5.x is incompatible with passlib 1.7.4

# ML/Data Science
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0
joblib==1.3.2

# Rate Limiting
slowapi==0.1.9

# Scheduling
apscheduler==3.10.4

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-dateutil==2.8.2

# Web Scraping
playwright==1.49.1
beautifulsoup4==4.12.3
